requirements.txt
tensorflow==2.18.0
flask
flask-cors
numpy
pillow